pyproject.toml
[project]
name = "siglip2-multimodal-hash"
version = "0.1.0"
description = "Multimodal image-text multi-label classification using SigLIP2, hashing, and KNN"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "TzuJeng, Lin (natsuki221)", email = "linnatsuki221@gmail.com"}
]

dependencies = [
    # Deep Learning Framework
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    
    # Transformers & Vision-Language Models
    "transformers>=4.40.0",
    "accelerate>=0.27.0",
    
    # Computer Vision
    "opencv-python>=4.9.0",
    "pillow>=10.2.0",
    "albumentations>=1.4.0",
    
    # Data Processing
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pycocotools>=2.0.7",
    
    # Similarity Search & Indexing
    # "faiss-gpu>=1.7.2",  # GPU 版本
    "faiss-cpu>=1.7.2",  # 如果沒有 GPU 則用此行
    
    # Metrics & Evaluation
    "scikit-learn>=1.4.0",
    "scipy>=1.12.0",
    
    # Visualization
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
    
    # Configuration & Logging
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.16.0",
    "tensorboard>=2.15.0",
    
    # Utilities
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    
    # Development Tools
    "ipython>=8.20.0",
    "jupyter>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py311', 'py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

# CUDA 12.4 版 PyTorch 官方 index
[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
# 強制 torch / torchvision / torchaudio 從 CUDA 12.4 index 取 wheel
torch = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux'" },
]
torchaudio = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux'" },
]

[tool.hatch.build.targets.wheel]
packages = ["src/siglip2_multimodal_hash"]

